requirements.txt
google-generativeai>=0.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
playwright>=1.40.0
markdown>=3.5.0
reportlab>=4.0.0
python-dotenv>=1.0.0
pydantic>=2.5.0
click>=8.1.0
aiohttp>=3.9.0
